see #459

See merge request !14

+++ src/main/java/com/wzsport/controller/SportScoreController.java
@@ -29,8 +29,8 @@ public class SportScoreController {
 	private SportScoreService sportScoreService;
 	
 	@RequestMapping(method = RequestMethod.POST)
-	public ResponseEntity<SportScore> save(@RequestParam int studentId,
-							 	 @RequestParam float meter50SprintTime,
+	public ResponseEntity<SportScore> save(@RequestParam long studentId,
+							 	 @RequestParam double meter50SprintTime,
 							 	 @RequestParam int meter50SprintScore,
 							 	 @RequestParam int standingJumpDistance,
 							 	 @RequestParam int standingJumpScore,
@@ -60,8 +60,8 @@ public class SportScoreController {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.POST)
 	public ResponseEntity<SportScore> update( 
-			 @PathVariable Integer id,
-			 @RequestParam(required = false) Float meter50SprintTime,
+			 @PathVariable long id,
+			 @RequestParam(required = false) Double meter50SprintTime,
 		 	 @RequestParam(required = false) Integer meter50SprintScore,
 		 	 @RequestParam(required = false) Integer standingJumpDistance,
 		 	 @RequestParam(required = false) Integer standingJumpScore,
